changes to net_if

+++ sys/net/include/net_if.h
@@ -495,7 +495,7 @@ int32_t net_if_get_pan_id(int if_id);
  *
  * @return  the PAN ID on success, -1 on failure.
  */
-int32_t net_if_set_pan_id(int if_id, uint16_t pan_id);
+int32_t net_if_set_pan_id(int if_id, uint32_t pan_id);
 
 /**
  * @}
